0. Print list.

+++ 0x12-singly_linked_lists/lists.h
@@ -12,9 +12,9 @@
 typedef struct list_s
 {
 char *str;
-size_t len;
+unsigned int len;
 struct list_s *next;
-}list_t;
+} list_t;
 int _putchar(char c);
 size_t print_list(const list_t *h);
 size_t list_len(const list_t *h);
